+++ NEWS.md
@@ -1,5 +1,8 @@
 # 0.4.0.9000
 
+* Changed `<=` and `>=` to `<` and `>`, respectively, to match app UI. Affects
+the functions `getResultsIntersection()`, `getEnrichmentsIntersection()`,
+`getResultsUpset()`, and `getEnrichmentsUpset()`.
 
 # 0.4.0